KO-384 Refactor out webhook
