+++ operator/cmd/manager/main.go
@@ -5,13 +5,9 @@ package main
 
 import (
 	"context"
-	"crypto/x509"
-	"encoding/base64"
-	"encoding/pem"
 	"errors"
 	"flag"
 	"fmt"
-	"io/ioutil"
 	"os"
 	"path/filepath"
 	"runtime"
@@ -33,12 +29,9 @@ import (
 	sdkVersion "github.com/operator-framework/operator-sdk/version"
 	"github.com/spf13/pflag"
 	v1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
-	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/apimachinery/pkg/util/intstr"
 	controllerRuntime "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/cache"
-	crclient "sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/client/config"
 	logf "sigs.k8s.io/controller-runtime/pkg/log"
 	"sigs.k8s.io/controller-runtime/pkg/manager"
@@ -275,109 +268,4 @@ func serveCRMetrics(cfg *rest.Config, operatorNs string) error {
 		return err
 	}
 	return nil
-}
-
-func ensureWebhookCertificate(cfg *rest.Config, namespace string) (err error) {
-	var contents []byte
-	if contents, err = ioutil.ReadFile(serverCertFile); err == nil && len(contents) > 0 {
-		certpool := x509.NewCertPool()
-		var block *pem.Block
-		if block, _ = pem.Decode(contents); err == nil && block != nil {
-			var cert *x509.Certificate
-			if cert, err = x509.ParseCertificate(block.Bytes); err == nil {
-				certpool.AddCert(cert)
-				log.Info("Attempting to validate operator CA")
-				verify_opts := x509.VerifyOptions{
-					DNSName: fmt.Sprintf("cassandradatacenter-webhook-service.%s.svc", namespace),
-					Roots:   certpool,
-				}
-				if _, err = cert.Verify(verify_opts); err == nil {
-					log.Info("Found valid certificate for webhook")
-					return nil
-				}
-			}
-		}
-	}
-	return updateSecretAndWebhook(cfg, namespace)
-}
-
-func updateSecretAndWebhook(cfg *rest.Config, namespace string) (err error) {
-	var key, cert string
-	var client crclient.Client
-	if key, cert, err = getNewCertAndKey(namespace); err == nil {
-		if client, err = crclient.New(cfg, crclient.Options{}); err == nil {
-			secret := &v1.Secret{}
-			err = client.Get(context.Background(), crclient.ObjectKey{
-				Namespace: namespace,
-				Name:      "cass-operator-webhook-config",
-			}, secret)
-			if err == nil {
-				secret.StringData = make(map[string]string)
-				secret.StringData["tls.key"] = key
-				secret.StringData["tls.crt"] = cert
-				if err = client.Update(context.Background(), secret); err == nil {
-					log.Info("TLS secret for webhook updated")
-					if err = ioutil.WriteFile(altServerCertFile, []byte(cert), 0600); err == nil {
-						if err = ioutil.WriteFile(altServerKeyFile, []byte(key), 0600); err == nil {
-							certDir = altCertDir
-							log.Info("TLS secret updated in pod mount")
-							return updateWebhook(client, cert, namespace)
-						}
-					}
-				}
-
-			}
-		}
-	}
-	log.Error(err, "Failed to update certificates")
-	return err
-}
-
-func updateWebhook(client crclient.Client, cert, namespace string) (err error) {
-	u := &unstructured.Unstructured{}
-	u.SetGroupVersionKind(schema.GroupVersionKind{
-		Group:   "admissionregistration.k8s.io",
-		Kind:    "ValidatingWebhookConfiguration",
-		Version: "v1beta1",
-	})
-	err = client.Get(context.Background(), crclient.ObjectKey{
-		Name: "cassandradatacenter-webhook-registration",
-	}, u)
-	if err == nil {
-		var webhook_slice []interface{}
-		var webhook map[string]interface{}
-		var ok, present bool
-		webhook_slice, present, err = unstructured.NestedSlice(u.Object, "webhooks")
-		webhook, ok = webhook_slice[0].(map[string]interface{})
-		if !ok || !present || err != nil {
-			log.Info(fmt.Sprintf("Error loading webhook for modification: %+v %+v %+v", ok, present, err))
-			return err
-		}
-		if err = unstructured.SetNestedField(webhook, namespace, "clientConfig", "service", "namespace"); err == nil {
-			if err = unstructured.SetNestedField(webhook, base64.StdEncoding.EncodeToString([]byte(cert)), "clientConfig", "caBundle"); err == nil {
-				webhook_slice[0] = webhook
-				if err = unstructured.SetNestedSlice(u.Object, webhook_slice, "webhooks"); err == nil {
-					err = client.Update(context.Background(), u)
-				}
-			}
-		}
-	}
-	return err
-}
-
-func ensureWebhookConfigVolume(cfg *rest.Config, namespace string) (err error) {
-	var pod *v1.Pod
-	var client crclient.Client
-	if client, err = crclient.New(cfg, crclient.Options{}); err == nil {
-		if pod, err = k8sutil.GetPod(context.Background(), client, namespace); err == nil {
-			for _, volume := range pod.Spec.Volumes {
-				if "cass-operator-certs-volume" == volume.Name {
-					return nil
-				}
-			}
-			log.Error(fmt.Errorf("Secrets volume not found, unable to start webhook"), "")
-			os.Exit(1)
-		}
-	}
-	return err
 }
